expand() filenames given to the ~@ tilde escape..
The ~r/~< and ~w tilde escapes yet expand()ed the user input,
whereas attachment filenames that were input via ~@ were not,
so that none of the ~+ expansions could be used.
So bring that up and enable the usual filename expansion
mechanism for both input forms of ~@.

Note:
arguments of the -a command-line option are not covered by this
changeset, since the necessary environment is not available once
those are evaluated.
(But fortunately the normal shell expansion is available on the
command line instead.)

The solution for this would be to use temporary storage for all
filenames given to -a arguments (it seems as if *argv* could be
reused for that, which is decently spaced sufficiently
automatically) and to perform all the expansions in a row once
everything is properly setup, but those changes are not covered
by this changeset.