Add *smime-sign-include-certs* (and *-user@host*) options..
Those can be used to specify additional certificates which are
to be included in the S/MIME message in addition to the main one
given in *smime-sign-cert*.
They are interpreted as lists of comma-separated filenames, each
of which containing a single certificate in PEM format.
The filenames are expanded as usual so that +/~ prefixes work.

This missing feature was discovered and analyzed by
Ezequiel Garzón (garzon (DOT) lucero (AT) gmail (DOT) com) and
Martin Neitzel (neitzel (AT) gaertner (DOT) de).
The author of the changeset was left alone in a horrible way in
the test phase.

P.S.:
the error execution paths in openssl.c could be made *much*
cleaner.  For now simply add yet another patch on top of it.