makeconfig, Makefile: extend to support variable INCLUDES..
The great plain automatic configuration system (i would never
dare to call that something with an Auto prefix though, how
could i) that Gunnar Ritter has written did yet not support the
addition of preprocessor conditionals, but only supports
conditional addition of libraries.

This is a pity since that disallows the automatic configuration
of GSSAPI under OpenBSD (and maybe others), since there all the
Kerberos stuff is under /usr/include/kerberosV!
This is a rather bizarre situation.
(I've also checked the OpenBSD nail package, and even that
doesn't patch it up.)

So extend the present infrastructure and allow the conditional
definition of something else than libraries.