makeconfig: add $tmp3, rework OpenBSD commit..
The commit (Enable the conditional detection of GSSAPI on
OpenBSD, 2012-05-23) adds a faulty redirection (i.e., yet
another uselessly escaped redirection terminator [the yet
present one was '%']), and duplicates the content of another
redirect.  Also the yet existent content of $tmp2.c is identical
to the mentioned two besides a single #include directive.

Extend makeconfig so that it supports a third temporary file
($tmp3), and use that instead of redirection for the mentioned
two places by simply using sed(1) on the content of $tmp2.c,
and not by recreating it from scratch.