Add -O option to pass options through to MTA..
Inspired by a mail from Dario Niedermann (dnied (AT) tiscali DOT it)
on the Heirloom mailx mailing list this changeset adds the
option -O which passes its arguments through to the chosen MTA,
but does not change the general POSIX usage syntax (remaining
arguments on the command line are solely interpreted as
recipient addresses).
Now it is possible to say

   $ nail -O-bm -O-C -O/etc/sendmail-alt.conf ..

to get the following sendmail command line.

   -i -bm -C /etc/sendmail-alt.conf