Allow for user-configured optional features..
Yet all features that can be supported on the operating system
are automatically enabled.  Being on the mailing list of
Heirloom mailx(1) (S-nail(1) ;) for about five months it seems
that almost everyone uses this program only as an automatic
mailer instead of interactive.
Also, the author of these lines does not make use of, e.g., IMAP.

Therefore it seems to be reasonable to introduce a new file
*user.conf* that can be used to adjust the set of features that
the configuration should search for.

This commit also fixes a faulty ID in the (disabled) use of
catgets(3) and adds the correct string to the catalog ("no IMAP
support" would yet have been translated to "no POP3 support").