Improve *quote-fold* behaviour..
(Implement *quote-fold* option.., 2012-08-27) failed to actually
merge multiple whitespaces in quotes, as in ">  > >   >".
Fix that.

Also extend the character classification functions with a
blankspacechar() functions and use that when testing for
whitespace in quotes.
blankchar() only tests blanks and whitechar() includes NL,
which is why i yet have not used the builtin mechanism.
Still there is a problem since U+00A0 NON-BREAKING-SPACE
is not covered.