Resort Makefile, allow easy name change..
Cleanup the Makefile a bit.
And introduce the new SID variable which can be used to specify
the S- prefix; i.e., with setting this to the empty string one
can enforce a name change to nail(1) all the way through.
(Well, except in the manual and such, but true regarding instal-
lation.)