INSTALL and README rework..
Rewind the changes and simply add S-nail(1) specifics on top
of the plain Heirloom original files.  Too bad it has not been
done like that from the beginning, fixing this up cannot be
done now that the repo is public.  :(
Also, due to that it is public, be more user friendly.