Avoid warnings for the bltin ctype stuff..
It yet checked that the character is valid ASCII before using it
as an array index, but still this produced compiler warnings.
So fix this, and make the implementation more eye friendly by
introducing generic implementations and call those from within
the real macros with only that stuff that matters.