Isolate iconv(3) usage in a single place..
Unfortunately the prototype differs in between systems.
POSIX wants char** for the constant *inb* argument, which is mysterious,
which is why some iconv(3) people actually changed this to the more
correct const char** (Citrus project, APR-iconv).
GNU libiconv even allows configuration-time changes (intransparently).

We cannot test this at configuration time since nail(1) goes for
complete POSIX embedding and relies on make(1) and the inference rules,
so that we cannot pass any -Werror to the compiler.

To avoid the warnings, use a simple preprocessor macro condition
and if it seems we know what to do, define a cast.