Allow full "RFC 5322 address"es for -b and -c..
(Fix what -r passes through to MTA.., 2012-09-18) allowed to
give a full RFC (said 5322 as it is the current) "address" as an
argument to -r, rather than only the plain "addr-spec", e.g.,
"Foo Bar <foo@bar.foo>" instead of "foo@bar.foo".
This commit adds this possibility for -b and -c.