Add ndup() function..
Yet a lot of string-to-struct name and struct name-to-string
convertions take place, multiple times for each address for
each mail that is composed.
Especially in respect to the upcoming IDNA support, but also
fundamentally, this is extremely wasteful and unnecessary.
Since the data is used in multiple lists and serves also
as a list node, introduce at least duplication of the prepared
raw data.