*Really* adapt Heirloom mailx(1) as S-nail(1)..
Since it is an official fork since my email to nail-devel@
on 2012-09-18, remove all occurrences of Heirloom mailx since
they are lies for the *master* branch.
Drop SCCS related stuff, too.

Also revise the version system by dropping the silly version.h
and replacing it with an empty version.c that gets occasionally
written by the Makefile on request (on a new release or so).

And drop all the packaging stuff Gunnar Ritter introduced.