Improve config/make system, silence smake(1)..
Yet Makefile was used for building of test programs as well as
for building of the actual program.
Surely this shows the power of inference rules and of the great
make(1) utility in general, but it results in nasty building
overhead due to the LIBS and INCS mechanism.

So let makeconfig create a temporary config.mk and use that
during the configuration phase.
One side effect of this is that makeconfig can now simply append
additional include paths and library specs to INCS/LIBS without
going through temporary files.
To ensure that INCS and LIBS consist of only one line a new
squeeze_em() function has been integrated into makeconfig
that'll use awk(1) to accomplish this after everything is well
set up.

See also this makeconfig change:

   -trap "rm -f $tmp.? $tmp ...
   +trap "rm -rf $tmp.*; rm -f $tmp ...

It happens that on Mac OS X some kind of BUILDNAME.dSYM
directories are created when the test objects are build, so
ensure those get thrown away.  (Haven't looked in detail, maybe
they are deleted automatically after a link-phase).

Because INCS and LIBS are now known to be oneliners once
Makefile uses them, cat(1) can be used instead of grep(1) to
integrate their content onto command lines.
($(< FILE) is not truly portable, unfortunately.)

Finally silence Jörg Schilling's smake(1) by adding .y to
.SUFFIXES and a dummy interference rule for .y to both
Makefiles.  It is a rather well-known bug of smake(1), but
adding the inference rule is better than looking at .. like

   smake: Recursion in default rule for 'edit.y.y.y.y.y.y.y.y'.