More on names and package friendliness..
So lets just go the whole way and ensure that forkers and/or
packagers have an easy job in the future and restrict the places
where adjustments are necessary to two: Makefile and nail.1.
And even offer the possibility to use a different name instead
of "nail", though the environment variables will remain NAILish.