About LINESIZE - and fix spurious BUFSIZ in favour of it..
LINESIZE should possibly be BUFSIZ, but for that the entire
program really needs to be carefully inspected to avoid
regressions.  However -- *do* make LINESIZE already an alias of
BUFSIZ *today* if the latter is greater than the former.

Also fix a spurious obvious misuse of BUFSIZ in collect.c.
That definitely should be LINESIZE.