Allow complete "to-addr"ss specs also for recipients..
The two commits (Allow full "RFC 5322 address"es for -b and
-c.., 2012-09-25) and (Fix what -r passes through to MTA..,
2012-09-18) introduced the possibility to add complete "address"
rather than only "addr-spec"ifications in RFC terms, i.e.,
including names and comments, for the mentioned options.

A user was however not yet able to do so for normal recipients
on the command line.
One was however always able to do so via aliases.
Finally change this for "to-addr"esses, too.