nail.1: drop of -h, correct docu for -O plus..
Fix one false macro invocation that has been introduced in
(s-nail.1: use \*(UA, \*(ua, \*(ba in place of NAILnames..,
2012-09-28).

-O was yet falsely documented not to take arguments (in
synopsis).  Do give an example and show how to pass a hop count
to old sendmails, now that -h has been finally dropped.

Do document what is accepted as addresses on the command line
now that full specifications may be passed.
And go the first of a few steps to restyle that messy manual.