Add WANT_IPV6 configuration support (default to =0)..
Offer IPv6 configuration support just as for all the other
optional features.
Since it was disabled in Heirloom mailx default to =0 by
default, though, unless i've had a chance to really look at the
code and test it, though.  Users that enabled it locally can now
however turn it on easily.