makeconfig: make compatible with old shells plus..
Yes there still was room for improvements.
Historic shells do not have "local", they do not like unary !not.
Also the wantfeat() i have introduced was unnecessary complicated,
which includes the explicit return of a value, so fix all that.

The squeeze_em() i have introduced to be able to simply cat(1)
the LIBS and INCS files printed an extra trailing space since
the awk(1) call did not reset the *ORS* variable.  Fixed.

Finally rely on inference rule when creating the config message ;),
instead of testing for a cc(1) and using that ;).
While there, do not mention Gunnar Ritter anymore, since the
grammatically correct mentioning of his name introduces an unmatched
single quote character that causes an error message.  Ouch!

Well and i have also started to introduce (redundant) quoting of
filenames but wasn't consequent.  Also change this, though it
must be said that all those quotes are in fact redundant. ??

Ah - one more.  Of course cat(1)ing a HERE document does not
perform variable expansion, so do get rid of the single one.