def.h: use 8 bit for char classification..
(Avoid warnings for the bltin ctype stuff.., 2012-09-22) changed the char
classification stuff and simply casted to an unsigned type (size_t),
which was in line with what was there (unsigned).
It is much better to treat the argument as an unsigned char instead
and however, so do change accordingly.
While there, properly enclose the macro arguments in parens.