Add userarg_extract() and use it for commline args..
The commits
(Allow complete "to-addr"ss specs also for recipients.., 2012-09-29),
(Allow full "RFC 5322 address"es for -b and -c.., 2012-09-25) and
(Fix what -r passes through to MTA.., 2012-09-18) used sextract(),
but can chooses which separators to use depending on content.
This is not what we need for user arguments for S-nail, since
an argument like "|cat > filename.txt" will be parsed as "|cat" and
"filename.txt".

So introduce a new userarg_extract() that always chooses the version
that requires a comma separated list.