sendout(): FIX, recode and implement *add-file-recipients*..
Use a flag carrier instead of variables.
Write the comma once it is clear that *this* name will be
written, not when we look at the *next* name and see that
this may be written.  This fixes faulty creations like

  echo bla | heirloom -s au t1@x.com ./tof1 ./tof2 t2@x.com

which creates

  To: t2@x.com t1@x.com

Also implement the new *add-file-recipients* option that can
be used to specify that file and pipe-command addressees shall
be included in the corresponding recipient list of the final
message.
Before files were always ex-, and pipes always included.