outof(): FIX and recode..
Yet it was impossible to share files and pipes as recipients,
or to use multiple pipe addressees.
In the former case only the files were served, in the latter
only one pipe got any data.
This behaviour was undocumented, though the code contained an XXX.
So reimplement this as such.

First check if there is any file or pipe recipient at all, and
simply perform a fast return if there is none.

Count the pipe addressees and create a descriptor array that'll
later be filled in with newly open(2)ed duplicates of the
original file, so as to avoid file-offset sharing in between
dup(2)ed fds.

Regardless of how many file recipients there are, only use one
reading stream instead of dup(2)ing and newly creating such
a stream for each recipient.  I failed to see why this should be
necessary since this is a sequential list walk.

If the mail message image can't be created don't retry to do so
but jump to a shorthand loop that only GDELs the file and pipe
recipients.

While there change the error messages to something useful.