Make system: create adjusted manual for installation++..
Create an adjusted manual dependend on user chosen Makefile
values.

Except for the .rc file it is now possible to create a clean
S-nail(1) fork simply by adjusting the Makefile (or even the
make(1) commline).

Once nice side effect is that the path of the installed
system-wide configuration file will always correspond to what
the binary expects...
Because the author was irritated first do change the MAILRC
variable that was used as a path to this file to SYSCONFRC,
which seems to be more correct (POSIX: "an unspecified system
start-up file").