*writebackedited* hack..
Sofar if the final newlines of an edited message has been
stripped away the resulting MBOX file was broken, because
the required separating newline was not automatically added
again, i.e., that was part of the edit action!

So do a 15 minutes quick and dirty hack to ensure that the
resulting MBOX is a valid one.

THIS MAY NOT BE CORRECT in that noone knows what may be the
target, and the better approach would be to only pass through
to the user what really can be edited, but due to the way
the code works this is not that easy.
It seems to work for normal MBOX input and normal MBOX output.

NetBSD mailx(1) suffers from the same bug.