Rename hfield() to hfieldX(), add hfield1()..
These are all macros that expand to invocations of hfield_mult().
While trying to track down buggy MBOX handling i have seen that
searching for a header field continues even if one has been seen
already and that header must be unique according to RFC 5322.

It turned out that this kind of cruel performance killing was
present all around, also for this issue.
So introduce the mentioned macros and do not continue searching
for headers that are allowed to occur once only.

While there drop support for a nonstandard Subj: abbreviation
for Subject: which is (a) not documented and (b) yes, i do have
something similar for my stale S-Postman project.  It's Python.