Tweak (Use system getopt(3) if available.., 2012-10-05)..
That commit had two pitfalls.

First it simply used getopt() if available, though getopt.c has
been introduced to be able to provide a non-reordered argument
list processing.  (And that was also needed because in main()
direct messing around with the argument list had been performed;
this is no longer true for S-nail(1), though.) So never succeed
and use libc getopt(3) when __GLIBC__ or __linux__ preprocessor
variables are detected.

Secondly the author introduced user.conf to express user
options, and these are then (when satisfieable) defined as
USE_*, rather than HAVE_*.  The mentioned commit used USE_*,
too, because the author planned to make usage of the builtin
getopt() configurable in user.conf first.  And, because
makeconfig link_check() requires a #define that it can set when
the check succeeds, that required usage of a pseudo variable
since USE_BUILTIN_GETOPT could of course not be used.  Well.
Instead of turning that to USE_SYSTEM_GETOPT (which would thus
have been wrong due to multiple reasons) turn this
non-user-configurable thing to HAVE_GETOPT.