newfile(): expand() part filename user input..
On our way to be more interactive and be more convenient for users,
offer usual folder name expansion also for the requested user input
during the *write* command.

Loop on expansion failure, too.  I first had a more complicated version
that also used yorn() on the expanded input, but i think that is
possibly overkill for a S-nail(1) user.  Though...