Straighten prefix handling..
A terrible mess with prefix handling is fixed with this one.
Before trailing whitespace was detected and the prefixlen was
adjusted accordingly -- but then not really used further on
(more exactly: sometimes yes, sometimes not; the author of this
changeset was not aware of this and simply used the passed prefixlen
to make decisions: how foolish!).

Change this.  Do not strip the trailing whitespace but simply
detect the length of the user prefix, and then really use that
all along the way instead of calculating the prefix length over
and over again, and, worse, assuming a NUL terminated prefix.

While there, take some more care about the statistics that were
not updated at all places.  This is still not a complete solution
to this problem.

And do a smallest-possible-step towards error handling, that is
yet simply absent.