Let some functions take "char const *" instead of "char *"..
The const keyword was already existent in C89, so the time has
come to slowly convert the code base and use it.  .)

Do so now for addrspec_with_guts() skip_comment(), unixtime()
and rfctime().

This changeset also does so for nexttoken(), though this is
a temporary change only, since this function will go away soon.

squash! Let some functions take "char const *" instead of "char *"..