Rewrite *extract()..
The *extract() series was "almost completely" broken yet.  If
a field with a leading comment was parsed, that was thrown away,
but a trailing one not -- unless the field started off with
angle brackets, in which case at the end of the closing angle
bracket parsing was finalized.  Also the comment parser did not
take care for backslash escaping, i.e., *quoted-pair*s according
to RFC.  And it did not take care for quotes at all, even those
need some special treatment, including the mentioned
*quoted-pair* handling.

So rewrite the extraction code.
Handle all of the above, and also squeeze adjacent whitespace to
single SP characters, as is prodded by RFC.

Note it still is not compliant to RFC 5322, see TODO for more.
We maybe should add angle brackets around addresses if missing
anyway, regardless of wether we follow the RFC in respect to
SHOULD NOT include comments.  In respect to that: i don't like
how mutt(1) simply converts comments to *phrase* (acc. 2 RFC).
(Not to talk about not correctly squeezing multiple WSP that
comes from mergin multiple comments.)

Note this finally also fixes an error with address boundary
detection in the input string.