Use more lextract() -- drop sextract()..
Maybe too optimistic in respect to backward compatibility, but
*addr-list* has always been comma separated.
The only list that were not are references, and these are served
well by extract().

So this leaves us with some misuses of lextract(), since that is
called for Sender: etc.;  but is this really worse than before,
once sextract() was used for those non-list field?  I think not.

While there, get rid of two useless temporary string allocations
in openssl.c.

squash! [improve-name-extraction] Use more lextract() were reasonable..

Drop sextract()!