expand(),file_expand(): use a shared implementation..
I have seen that a greek colossa actually grumbled that file_expand()
does not expand the decade-known % and such, but only + and ~.
So do share the implementation in between them and *do* expand the
mentioned, yet fail for file_expand() if the result is not local.

It makes the file_expand() case needlessly slow, but this is
a grown historic program and users should generally be able to
do whatever they want anyway.