Do not warn when handling -S, the first time..
(main.c: set -S twice!  And about command line processing..,
2012-10-18)..  fixed the problem that -S settings were not
active during further command line processing by set()ing the
user desire twice.

However, if the user effectively initiates an unset request it
most likely results in (at least) a warning of the option is not
set.

To overcome this use the unset_allow_undefined= mechanism to
suppress the warning for the first set().  If the second set()
that happens after the resource files are loaded still does not
find anything to unset the warning will be fine (maybe).