Do not use non-compiler-builtin alloca(3)s..
Due to NetBSD PR lib/47120 it seems to be much safer to not
use non-compiler-builtin versions of alloca(3), since modern
compilers just cannot be trusted not to overoptimize and
silently break code.

So test and go for __builtin_alloca() only, and let def.h
use HAVE_ALLOCA as an expansion for the actual cc function,
just in case, and for future safety.

While there it was detected that the (now optional) builtin
getopt implementation messed around and defined alloca stuff
on its own; change that and simply include rcv.h, which also
includes all other headers that are desired.