FIX MIME quoted-printable encoding (char cast)..
(;-{
Auauauauau!!

Well, one of the things that have already been started in v13 is
the turn from using "int" when working with 8-bit characters to
"unsigned char" (and as long as we do not support wide
characters).

Unfortunately one very important piece of code, that is handling
encoding to quoted-printable, still used integer instead of
unsigned char, which caused an automatic extension cast to take
place, and that resulted in a messed up output.
Sorry!