smtp_mta() only if USE_SMTP, correct start_mta()..
(Improve sendmail(1) argument processing.., 2012-10-22) and, indeed,
(sendout.c: *expand() may fail, 2012-10-23) introduced an "error" in
that the compiler may complain about "mta might be used uninit", but
that is wrong.  But i do not in fact want to follow this rather usual
programming tradition, so just fix it!

While there again, drop the smtp_mta() hook completely if USE_SMTP is
not defined.  One of the goals of S-nail(1) is to drop all function
stubs that only exist because of lazy programming at first hand,
so go a small step in this direction.  (Too late though.)