Cross-list remove recipient duplicates..
Get rid of one TODO entry and of one nail "bug" that will not be
backported to heirloom-plus.

Yet, if aliases would expand to multiple identical recipients,
as can happen, e.g., if ~/.mail_aliases is included multiple
times, then the recipient fields will show duplicates, whereas
the actual addresses for sendmail(1) or SMTP invocations would
have those duplicates eliminated already.

It turns out that solving this problems *seems* to fix a TODO
entry, i.e., that recipients that are in To: should not also
occur in Cc:, and ditto...

The implementation still is terribly inefficient and needs a lot
of more work to be done.