Strip Content-length: plus (add *keep-content-length*)..
We do neither use nor manage Content-Length: and Lines:, and
after modifying a message those fields are most likely plain
wrong.  It seems best to discard them after an update.

Since not all users may like this behaviour, add
a *keep-content-length* variable that revives it.

But maybe we should not be configurable.
mutt(1) adds them silently and completely on its own.