Implement RFC 4155 conforming MBOX parsing..
The actual code is a heavily edited mix of what was already
there, but disabled, and NetBSDs Mail(1).
This changeset does not yet change the way that "From " lines
are quoted.

While there adjust fakefrom() to return "MAILER-DAEMON" as
described in some old mbox manual, instead of the (just as
non-RFC 4155-compliant) plain hyphen "-".