Fix MTA/*record* descriptor clash..
Ooooh, for two months MTA interaction was broken when *record*
was also sent!
Introduced in (If *record* is set, avoid writing dead content
twice.., 2012-09-14), before i really knew what i was doing i'm
afraid, and you have to for this codebase.

This change in fact caused the same "descriptor offset shared by
child" problem that i've fixed for pipe and file recipients in
(outof(): FIX and recode.., 2012-10-04) (and tweaks thereof).

Anyway, (ISO C and) POSIX don't define a real way to make
a FILE* revisit the *real* file offset of its underlaying
descriptor (fflush();rewind() won't do, or *BSD are all
broken), so the actual fix is rather hackish (first SEEK_END to
EOF, then rewind()).

In fact it must be said that the codebase would win a lot if we
would turn from FILE* to a buffer-based S-nail-specific file
object that would be unlocked and work directly on the
descriptors.  But that would be a lot of work.