Base64: implement a new b64_* fun family..
The current Base64 mechanism has weaknesses.
First it is overly complicated and inefficient.
Second it is broken, since it uses function-local statics to
keep state in between calls.
(This "mechanism" is also used in at least prefixwrite(), but it
surely is one of the most terrible things i've ever seen, in C.)

It breaks if a base64-encoded part is parsed that is invalid,
since some state is carried over to the next invocation, as i've
seen today (it actually made me start this topic branch), in
which case even binary content is passed through:

   ? p
   Message 21:
   From S-Postman Sat Nov 17 12:52:25 2012
   Date: Fri, 16 Nov 2012 20:00:09 -0800 (PST)
   To: daily-article-l@lists.wikimedia.org
   From: English Wikipedia Article of the Day
           <daily-article-l@lists.wikimedia.org>
   Subject: [Daily article] November 17: Metroid Prime
   Reply-To: dal-feedback@wikimedia.org

   �5��ɽ���Aɥ����́��٥��������[lots of data follows]

So turn and implement a new family of functions, that used to be
clean, efficient and error-tolerant.
During the life of this topic branch, replace all uses of the
old interface with the new one, then throw away that crap.