Tweak MIME boundary detection..
Counting separating newlines is a problem all through the
software (Berkeley Mail that is), and it also bites for message
boundary detection, as has been seen for a mail on a NetBSD list
that looks like that in mutt(1):

  I     1 <no description>          [text/plain, 8bit, iso-8859-1, 0.7K]
  I     2 <no description>          [multipa/appledoubl, 7bit, 3.9K]
  A     3 |- %areca-1882.diffs      [applica/applefile, base64, 1.9K]
  A     4 |- areca-1882.diffs       [applica/octet-stre, base64, 1.4K]
  I     5 <no description>          [multipa/appledoubl, 7bit, 16K]
  A     6 |- %NetBSD-6-dmesg.txt    [applica/applefile, base64, 2.0K]
  A     7 |- NetBSD-6-dmesg.txt     [text/plain, 7bit, us-ascii, 14K]

The problem was the missing empty line in between boundaries.
A broken down test mail looks like (output stripped) that in
Heirloom mailx(1):

   Part 1:
   Part 2:
   Part 2.1:
   Part 2.2:

   --============_-857507304==_D============--
   Part 3:
   Part 3.1:
   Part 3.2:

   --============_-857507304==_E============--

While here, change mime_getboundary() prototype to store the
already calculated length in a new argument, if provided.