Change "Part X:" display message..
Take care of wether the Content-type: is ignored or not,
and do print some more info if it is.
Also place the entire message in brackets, so that it
is easier to separate at a glance, imo.
This can be a first step only, since all these infos
could very well be compressed if the part does not
contain any data etc., i.e., the separator line algorithm
is pretty much nonexistent.  At some future time this
is the desired display:

 [Part XY - TYPE, SIZE[, FILENAME]]
 [Part XY ...]

I.e., no separators unless content is displayed, attachment
filenames if present should be displayed, and these lines
should very well fit on the users screen, of course.