Introduce global *verbose*, use it..
Sofar many files used local *debug* and/or *verbose* variables
to track the *current* state of the very variables.
This is sick.

Introduce a global *verbose* variable in equal spirit to the
global *debug*, and use that everywhere.  Let *verbose* be
implied by *debug*.

Adjust assign() and unset_internal() so that they look for
special variables (yet debug and verbose, say), and perform
updates if necessary.
(This may be improved by a generic implementation once the value
cached is implemented, but i'm unsure yet.)