schdir(): realpath() local files before leaving CWD..
S-nail (and all BSD Mails afaik) has the problem that the
program effectively locks up if the currently open folder cannot
be accessed no more.

Implement a short workaround for the "chdir" command that will
first try to create an absolute path for the folder, and refuse
to perform the chdir if that fails.
This workaround requires realpath(3).

It is not a solution for the actual problem, since that arises
in a lot of different conditions (like concurrent removal etc.
of the mailbox), but the real solution will need some more years
of development, and this simply commit gives a hand *now*.