Introduce *charset-8bit* and *charset-7bit* variables..
(Remove use of undocumented *charset7* option.., 2012-12-21)
was committed and that was the right thing to do.

However, in the current codebase, we fall back to a seven bit
charset if the MIME classification does not find any high bit
set, and this won't work if the charset is *not* compatible with
US-ASCII.  And besides, since we have
a Content-Transfer-Encoding: header field and correctly use it,
one should be able to say LATIN1.  Or even UTF-8.

This is even more true for the default outgoing character set
that is used if *sendcharsets* is not set (i.e., *charset-8bit*).

All of this is intermediate anyway.
After the send and MIME layer rewrite that will cumulate as
v15.0 we will have input character sets and output character
sets, and then there is a terminal character set, too.
See TODO (of v14.0) for more on this.