MIME: add charset_iter_*() family..
Right now we use a global *wantcharset* variable and mess around
with it from multiple places in the code.  Introduce a generic
approach that will provide the necessary and/or possible
charsets in correct order.

It yet offers legacy functions to support iterator recursion,
but that will no longer be necessary after the send/MIME layer
rewrite that'll cumulate in v15.0 (as that'll write attachments
sequentially into a single temporary file, and *then* the text
part into the final message temporary, and then integrate the
attachments in reverse order [truncation possible] into the
message.)