collect: support multiple attachment charsets..
Ooops!  When i hurried over fixing compiler warnings the first
time, back in september (...), i've introduced a bug that
prevented asking for character sets of attachments since then.
Introduced in (collect.c: fix compiler warnings.., 2012-09-24),
fixed today.

But hey - this shit was crap.  Get rid of these BANANAS!

In the future we offer multiple ways how to treat attachments,
including specifying input character sets that are passed
through unchanged, a specific conversion specified input charset
-> specified output charset etc.

This will be even better after the MIME and send layer rewrite
when there will be only one temporary file regardless of the
amount of attachments and what the user does with 'em, and of
course we will need command line arguments that offer the
possibility to specify the conversions, too.